T20-cricketdataanaylsis-main/T20-cricketdataanaylsis-main/requirements.txt
deta==1.1.0
gitdb==4.0.10
GitPython==3.1.31
httpx-oauth==0.11.2
iconify==0.0.103
idna==2.10
importlib-metadata==6.1.0
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
plotly-express==0.4.1
streamlit==1.21.0
streamlit-authenticator==0.1.5
streamlit-option-menu==0.3.2
XlsxWriter==3.1.0
python-dotenv==1.0.0
